Sale : * Merge files * delete already merged file
